refactoring after code review (#1, #2)
- move package from `/badge` directory to the library root
- add LICENSE
- remove `vendor/` from git
- add some comments
- get rid of the `sync.Once` for template initialisation
- inline content of the `init` function
- extract `map[string]interface{}` template's data to internal badge `struct`
