+++ badge_test.go
