+++ fonts/Bitstream Vera License.txt
